[master]: task N4 Description At this stage, your calculator should support the addition + and subtraction - operators.
The program must calculate expressions like these: 4 + 6 - 8, 2 - 3 - 4, and so on. It must support both unary and binary minus operators. If the user has entered several same operators following each other, the program still should work (like Java or Python REPL).

Consider that the even number of minuses gives a plus, and the odd number of minuses gives a minus! Look at it this way: 2 -- 2 equals 2 - (-2) equals 2 + 2.

Modify the result of the /help command to explain these operations.

Decompose your program using methods to make it easy to understand and edit later.